# swagger-service/public/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/events": {
            "get": {
                "description": "Get a paginated list of events with optional filtering by user_id, item_id, event_type, date range, etc.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data-management"
                ],
                "summary": "List events with pagination and filtering",
                "operationId": "listEvents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Namespace",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit (default: 100, max: 1000)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset (default: 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by user ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by item ID",
                        "name": "item_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by event type",
                        "name": "event_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by creation date (ISO8601)",
                        "name": "created_after",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by creation date (ISO8601)",
                        "name": "created_before",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    }
                }
            }
        },
...
        "/v1/items/{item_id}/similar": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ranking"
                ],
                "summary": "Get similar items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "item_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "default",
                        "description": "Namespace",
                        "name": "namespace",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Top-K",
                        "name": "k",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/specs_types.ScoredItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    }
                }
            }
        },
...
        "/v1/recommendations": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ranking"
                ],
                "summary": "Get recommendations for a user",
                "parameters": [
                    {
                        "description": "Recommend request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.RecommendRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.RecommendResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.APIError": {
            "type": "object",
            "properties": {
                "correlation_id": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "specs_types.ScoredItem": {
            "type": "object",
            "properties": {
                "explain": {
                    "$ref": "#/definitions/types.ExplainBlock"
                },
                "item_id": {
                    "type": "string",
                    "example": "i_101"
                },
                "reasons": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "score": {
                    "type": "number",
                    "example": 0.87
                }
            }
        },
...
        "types.ExplainBlend": {
            "type": "object",
            "properties": {
                "alpha": {
                    "type": "number"
                },
                "beta": {
                    "type": "number"
                },
                "contrib": {
                    "$ref": "#/definitions/types.ExplainBlendContribution"
                },
                "cooc_norm": {
                    "type": "number"
                },
                "emb_norm": {
                    "type": "number"
                },
                "gamma": {
                    "type": "number"
                },
                "pop_norm": {
                    "type": "number"
                },
                "raw": {
                    "$ref": "#/definitions/types.ExplainBlendRaw"
                }
            }
        },
        "types.ExplainBlendContribution": {
            "type": "object",
            "properties": {
                "cooc": {
                    "type": "number"
                },
                "emb": {
                    "type": "number"
                },
                "pop": {
                    "type": "number"
                }
            }
        },
        "types.ExplainBlendRaw": {
            "type": "object",
            "properties": {
                "cooc": {
                    "type": "number"
                },
                "emb": {
                    "type": "number"
                },
                "pop": {
                    "type": "number"
                }
            }
        },
        "types.ExplainBlock": {
            "type": "object",
            "properties": {
                "anchors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "blend": {
                    "$ref": "#/definitions/types.ExplainBlend"
                },
                "caps": {
                    "$ref": "#/definitions/types.ExplainCaps"
                },
                "mmr": {
                    "$ref": "#/definitions/types.ExplainMMR"
                },
                "personalization": {
                    "$ref": "#/definitions/types.ExplainPersonalization"
                }
            }
        },
        "types.ExplainCapUsage": {
            "type": "object",
            "properties": {
                "applied": {
                    "type": "boolean"
                },
                "count": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "types.ExplainCaps": {
            "type": "object",
            "properties": {
                "brand": {
                    "$ref": "#/definitions/types.ExplainCapUsage"
                },
                "category": {
                    "$ref": "#/definitions/types.ExplainCapUsage"
                }
            }
        },
...
        "types.ExplainMMR": {
            "type": "object",
            "properties": {
                "lambda": {
                    "type": "number"
                },
                "max_sim": {
                    "type": "number"
                },
                "penalty": {
                    "type": "number"
                },
                "rank": {
                    "type": "integer"
                },
                "relevance": {
                    "type": "number"
                }
            }
        },
        "types.ExplainPersonalization": {
            "type": "object",
            "properties": {
                "boost_multiplier": {
                    "type": "number"
                },
                "overlap": {
                    "type": "number"
                },
                "raw": {
                    "$ref": "#/definitions/types.ExplainPersonalizationRaw"
                }
            }
        },
        "types.ExplainPersonalizationRaw": {
            "type": "object",
            "properties": {
                "profile_boost": {
                    "type": "number"
                }
            }
        },
...
        "types.ListResponse": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {}
                },
                "limit": {
                    "type": "integer"
                },
                "next_offset": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "types.Overrides": {
            "type": "object",
            "properties": {
                "bandit_algo": {
                    "type": "string"
                },
                "blend_alpha": {
                    "type": "number"
                },
                "blend_beta": {
                    "type": "number"
                },
                "blend_gamma": {
                    "type": "number"
                },
                "brand_cap": {
                    "type": "integer"
                },
                "category_cap": {
                    "type": "integer"
                },
                "covis_window_days": {
                    "type": "integer"
                },
                "mmr_lambda": {
                    "type": "number"
                },
                "popularity_fanout": {
                    "type": "integer"
                },
                "popularity_halflife_days": {
                    "type": "integer"
                },
                "profile_boost": {
                    "type": "number"
                },
                "profile_top_n": {
                    "type": "integer"
                },
                "profile_window_days": {
                    "type": "integer"
                },
                "purchased_window_days": {
                    "type": "integer"
                },
                "rule_exclude_events": {
                    "type": "boolean"
                }
            }
        },
        "types.RecommendBlend": {
            "type": "object",
            "properties": {
                "als": {
                    "type": "number",
                    "example": 0
                },
                "cooc": {
                    "type": "number",
                    "example": 0.7
                },
                "pop": {
                    "type": "number",
                    "example": 0.3
                }
            }
        },
        "types.RecommendConstraints": {
            "type": "object",
            "properties": {
                "created_after": {
                    "description": "Optional ISO8601 timestamp; only consider events on/after this instant.",
                    "type": "string",
                    "example": "2025-06-01T00:00:00Z"
                },
                "exclude_item_ids": {
                    "description": "Exclude these item IDs from results.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "include_tags_any": {
                    "description": "Match if item.tags overlaps these (any). Empty/omitted = no tag filter.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "price_between": {
                    "description": "Optional price bounds: [min, max]. Either end may be omitted.",
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                }
            }
        },
        "types.RecommendRequest": {
            "type": "object",
            "properties": {
                "blend": {
                    "$ref": "#/definitions/types.RecommendBlend"
                },
                "constraints": {
                    "$ref": "#/definitions/types.RecommendConstraints"
                },
                "context": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "explain_level": {
                    "type": "string",
                    "enum": [
                        "tags",
                        "numeric",
                        "full"
                    ],
                    "example": "numeric"
                },
                "include_reasons": {
                    "type": "boolean",
                    "example": true
                },
                "k": {
                    "type": "integer",
                    "example": 20
                },
                "namespace": {
                    "type": "string",
                    "example": "default"
                },
                "overrides": {
                    "$ref": "#/definitions/types.Overrides"
                },
                "user_id": {
                    "type": "string",
                    "example": "u_123"
                }
            }
        },
        "types.RecommendResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/specs_types.ScoredItem"
                    }
                },
                "model_version": {
                    "type": "string",
                    "example": "pop_2025-09-07_01"
                },
                "profile_id": {
                    "type": "string",
                    "example": "vip-high-novelty"
                },
                "segment_id": {
                    "type": "string",
                    "example": "vip"
                }
            }
        },
...
    }
}